refact: add missing 'orca' package to :platform:starter
