+++ platform/starter/lifecycle/src/main/java/com/jeanbarrossilva/orca/platform/starter/lifecycle/CompleteLifecycleActivity.kt
@@ -20,7 +20,7 @@ import androidx.lifecycle.Lifecycle
 import com.jeanbarrossilva.orca.platform.starter.lifecycle.state.CompleteLifecycleState
 import com.jeanbarrossilva.orca.platform.starter.lifecycle.state.compareTo
 import com.jeanbarrossilva.orca.platform.starter.lifecycle.state.next
-import com.jeanbarrossilva.platform.starter.StartableActivity
+import com.jeanbarrossilva.platform.orca.starter.StartableActivity
 
 /**
  * [StartableActivity] that provides a complete version of [Lifecycle.State], containing missing
