+++ platform/starter/src/androidTest/java/com/jeanbarrossilva/orca/platform/starter/ActivityStarterTests.kt
@@ -13,9 +13,11 @@
  * not, see https://www.gnu.org/licenses.
  */
 
-package com.jeanbarrossilva.platform.starter
+package com.jeanbarrossilva.orca.platform.starter
 
 import androidx.test.platform.app.InstrumentationRegistry
+import com.jeanbarrossilva.platform.orca.starter.StartableActivity
+import com.jeanbarrossilva.platform.orca.starter.on
 import io.mockk.spyk
 import io.mockk.verify
 import kotlin.coroutines.resume
